java/src/test/java/com/yourtion/leetcode/daily/m09/d01/SolutionTest.java
package com.yourtion.leetcode.daily.m09.d01;

import com.yourtion.leetcode.TestUtils;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.params.provider.Arguments.arguments;

@DisplayName("每日一题 - 20200901")
class SolutionTest {

    static Stream<Arguments> testDataProvider() {
        return Stream.of(
                arguments("[1, 5, 2]", false),
                arguments("[1, 5, 233, 7]", true)
        );
    }

    @ParameterizedTest()
    @MethodSource("testDataProvider")
    void predictTheWinner(String source, boolean res) {
        System.out.printf("runTest: %s , res: %b", source, res);
        int[] ss = TestUtils.stringToIntegerArray(source);
        Assertions.assertEquals(res, new Solution().PredictTheWinner(ss));
    }
}